Fix scss issue when horizon is not run on the root of a url.
Because the scss file would need to have a prepended base url + only 1
small screen was using it + mainly because the fact that this scss file might
not get compressed properly on production systems. For now, the nuage.scss
file is removed. And the required css is moved to the 1 template that is
using it.

Change-Id: If7c4c93bf220096270d8269d4e37196ec31e6a41
Closes-Bug: OPENSTACK-1462